+++ source/general/registry.h
@@ -24,7 +24,7 @@ public:
 
 private:
     static SimulationPropertiesStorage m_SimulationProperties;
-    static float m_DeltaTime;
+    static double m_DeltaTime;
     static std::unique_ptr<Window> m_MainWindow;
     static std::unordered_map<std::string,std::unique_ptr<Window>> m_SubWindows;
